Created Functions
Yeah. Added functions to clean up a lot of the image embed commands. Added another usable prefix. Added some more NSFW commands and SFW commands.

Code should be less of a headache to read now

Also, it's not playing a game anymore; it's twitch streaming. And yes, the twitch channel is legit.