Created Refined Code
New code is now 100% more efficient (that's a lie, it's just much much shorter and still does most of the same commands) Refined Code so much better now. Like, it's only 83 lines long. Compared to the 250-something lines on the original. Please, please use the refined version. I can't be bothered to remove the old one since it acts as a reminder now on how  dumb it was to not use functions and the command extension for discord thinking it was waay to hard to learn when it literally took me half the time to learn it and make it work.